Fix no-any-return errors (48 fixes): add explicit type casts for torch.Tensor.item() and dict returns
Fixed 48+ no-any-return mypy errors by adding explicit type casts:

- torch.Tensor.item()  float() cast (decay factors, rates, values)
- dict operations  dict() or Dict[str, Any] cast
- bool comparisons  bool() cast
- Optional[int] assignments  explicit type annotation
- bytes operations  bytes() cast (compression)
- List/Episode returns  explicit type annotation

Files modified:
- mixins/device_mixin.py: Cast device type and cuda check
- diagnostics/: Cast criticality, performance metrics
- components/: Cast neuron properties, STP decay factors
- learning/: Cast BCM, metabolic, EI balance, social learning
- regions/: Cast striatum, hippocampus, prefrontal returns
- io/: Cast compression, encoding, checkpoint operations
- memory/: Cast consolidation similarity scores

Progress: 66  36 remaining no-any-return errors

Note: Black auto-formatted 2 files (neuron_factory.py, prefrontal.py)

+++ src/thalia/components/coding/spike_utils.py
@@ -50,7 +50,7 @@ def compute_firing_rate(spikes: torch.Tensor) -> float:
     """
     if spikes.numel() == 0:
         return 0.0
-    return spikes.float().mean().item()
+    return float(spikes.float().mean().item())
 
 
 def compute_spike_count(spikes: torch.Tensor) -> int:
